+++ src/FinalProject/model/Transaction.java
@@ -2,9 +2,9 @@ package FinalProject.model;
 
 public class Transaction {
     public String from, to, fileName;
-    public int amount;
+    public double amount;
 
-    public Transaction(String from, String to, int amount, String fileName) {
+    public Transaction(String from, String to, double amount, String fileName) {
         this.from = from;
         this.to = to;
         this.amount = amount;
